[TM-3294] change context instead of boolean for different maps
