+++ src/components/elements/Map-mapbox/Map.d.ts
@@ -80,7 +80,8 @@ export type PopupComponentProps = {
   setFilters?: DashboardFilters;
   dashboardCountries?: CountriesProps[];
   dashboardMode?: string;
-  newStyling?: boolean;
+  /** When true, map uses the champions (non-admin) layout and controls. */
+  championsMap?: boolean;
 };
 
 /** Data shape passed to setMobilePopupData; mirrors PopupComponentProps minus the live Popup instance. */
