+++ src/components/elements/Map-mapbox/Map.tsx
@@ -23,6 +23,7 @@ import { SitePolygonLightDto } from "@/generated/v3/researchService/researchServ
 import { useOnMount } from "@/hooks/useOnMount";
 
 import { addOrUpdateMarkerAndZoom } from "./adapters/camera";
+import { ChampionsMapProvider } from "./championsMap.context";
 import MapCanvas from "./components/MapCanvas";
 import MapControlsOverlay from "./components/MapControlsOverlay";
 import { PopupMobile } from "./components/PopupMobile";
@@ -80,8 +81,8 @@ export interface BaseMapProps {
   initialTileVersion?: string;
   /** When it matches current polygon data, skip bumping the tile cache on mount. */
   initialPolygonFingerprint?: string;
-  /** new Styling for polygons */
-  newStyling?: boolean;
+  /** Champions (non-admin) map layout and controls; omit or false for the default map. */
+  championsMap?: boolean;
 }
 
 export interface DashboardMapExtras {
@@ -142,7 +143,11 @@ export const MapEditingContext = createContext({
   setIsEditing: (_value: boolean) => {}
 });
 
-export const MapContainer = ({
+type MapContainerInnerProps = Omit<MapProps, "championsMap"> & {
+  mapFunctions: NonNullable<MapProps["mapFunctions"]>;
+};
+
+const MapContainerInner = ({
   onError: _onError,
   editable,
   geojson,
@@ -178,11 +183,8 @@ export const MapContainer = ({
   hasAccess,
   dashboardContext,
   disabledPolygonPanel = false,
-  newStyling = false,
   ...props
-}: MapProps) => {
-  if (mapFunctions == null) return null;
-
+}: MapContainerInnerProps) => {
   const { map, mapContainer, draw, onCancel, initMap } = mapFunctions;
 
   const {
@@ -344,8 +346,7 @@ export const MapContainer = ({
     setEditPolygon,
     editPolygon,
     setMobilePopupData,
-    dashboardContext: resolvedDashboardContext,
-    newStyling: newStyling
+    dashboardContext: resolvedDashboardContext
   });
 
   useMapCamera({
@@ -429,7 +430,6 @@ export const MapContainer = ({
     <MapEditingContext.Provider value={{ isEditing, setIsEditing }}>
       <MapCanvas mapContainer={mapContainer} className={className}>
         <MapControlsOverlay
-          newStyling={newStyling}
           hasControls={hasControls}
           draw={{
             handleEditPolygon,
@@ -460,13 +460,12 @@ export const MapContainer = ({
 
         {showLegend ? (
           <ControlGroup
-            newStyling={newStyling}
             position={
               disabledPolygonPanel ? "bottom-left" : siteData ? "bottom-left-site" : legendPosition ?? "bottom-left"
             }
             isFullscreen={isFullscreen}
           >
-            <FilterControl newStyling={newStyling} />
+            <FilterControl />
           </ControlGroup>
         ) : null}
 
@@ -487,6 +486,15 @@ export const MapContainer = ({
         ) : null}
       </MapCanvas>
     </MapEditingContext.Provider>
+  );
+};
+
+export const MapContainer = ({ mapFunctions, championsMap = false, ...rest }: MapProps) => {
+  if (mapFunctions == null) return null;
+  return (
+    <ChampionsMapProvider championsMap={championsMap}>
+      <MapContainerInner mapFunctions={mapFunctions} {...rest} />
+    </ChampionsMapProvider>
   );
 };
