+++ src/components/elements/Map-mapbox/components/ControlGroup.tsx
@@ -2,6 +2,8 @@ import classNames from "classnames";
 import { DetailedHTMLProps, HTMLAttributes, PropsWithChildren } from "react";
 import { twMerge } from "tailwind-merge";
 
+import { useChampionsMap } from "../championsMap.context";
+
 export type ControlMapPosition =
   | "top-right"
   | "top-left"
@@ -18,23 +20,24 @@ export interface ControlGroupProps
     PropsWithChildren {
   position: ControlMapPosition;
   isFullscreen?: boolean;
-  newStyling?: boolean;
 }
 
-const ControlGroup = ({ children, position, className, isFullscreen, newStyling, ...props }: ControlGroupProps) => {
+const ControlGroup = ({ children, position, className, isFullscreen, ...props }: ControlGroupProps) => {
+  const championsMap = useChampionsMap();
+
   return (
     <div
       {...props}
       className={twMerge(
         classNames("absolute z-[18] flex flex-col gap-8", {
-          "top-5": position.includes("top") && !newStyling,
-          "top-2": position.includes("top") && newStyling,
-          "left-5 items-start": position.includes("left") && !newStyling,
-          "left-2 items-start": position.includes("left") && newStyling,
-          "right-5 items-end": position.includes("right") && !newStyling,
-          "right-2 items-end": position.includes("right") && newStyling,
-          "bottom-8": position.includes("bottom") && !newStyling,
-          "bottom-2": position.includes("bottom") && newStyling,
+          "top-5": position.includes("top") && !championsMap,
+          "top-2": position.includes("top") && championsMap,
+          "left-5 items-start": position.includes("left") && !championsMap,
+          "left-2 items-start": position.includes("left") && championsMap,
+          "right-5 items-end": position.includes("right") && !championsMap,
+          "right-2 items-end": position.includes("right") && championsMap,
+          "bottom-8": position.includes("bottom") && !championsMap,
+          "bottom-2": position.includes("bottom") && championsMap,
           "!left-[30vw] items-start": position.includes("site"),
           "!left-5": isFullscreen,
           "left-[45%]": position.includes("center"),
