+++ src/pages/site/[uuid]/index.page.tsx
@@ -46,7 +46,7 @@ const SiteDetailPage = () => {
   const TabItems = [
     { key: "overview", title: t("Overview"), body: <SiteOverviewTab site={site!} refetch={refetch} /> },
     { key: "details", title: t("Site Details"), body: <SiteDetailTab site={site!} /> },
-    { key: "polygons", title: t("Polygons"), body: <SitePolygonsTab site={site!} refetch={refetch} /> },
+    { key: "polygons", title: t("Polygons"), body: <SitePolygonsTab site={site!} /> },
     { key: "map", title: t("Site Map"), body: <SiteMapTab site={site!} refetch={refetch} /> },
     {
       key: "gallery",
